context_management/pyproject.toml
[project]
name = "context-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "mypy>=1.16.0",
]
